+++ Arena/ArenaHooks.cs
@@ -184,6 +184,7 @@ namespace RainMeadow
             IL.ArenaGameSession.ctor += ArenaGameSession_ctor_IL;
             new Hook(typeof(ArenaSetup.GameTypeSetup).GetProperty("ScoreToEnterDen").GetGetMethod(), this.ScoreToEnterDen);
             IL.HUD.PlayerSpecificMultiplayerHud.Update += PlayerSpecificOnlineHud_Update;
+            DrownHooks();
         }
 
         public void PlayerSpecificOnlineHud_Update(ILContext il)
@@ -488,6 +489,7 @@ namespace RainMeadow
                 timeUntilFadeout = arena.amoebaDuration * 40,
             };
             voidSpawn.behavior = new VoidSpawn.ChasePlayer(voidSpawn, room);
+            voidSpawn.swimSpeed = 1f;
             room.abstractRoom.AddEntity(apo);
             RainMeadow.sSpawningNonTransferable = false;
 
@@ -553,6 +555,7 @@ namespace RainMeadow
             VoidSpawn.ChasePlayer self
         )
         {
+            
             if (!isArenaMode(out var arena))
                 return orig(self);
             //only runs on the person who created the voidspawn because voidspawn.behaviour is null on default and isnt synced
@@ -986,7 +989,7 @@ namespace RainMeadow
                 {
                     if (OnlineManager.lobby.isOwner)
                     {
-                        arena.SetPlayerStatsFromLocalPlayer(player, pl);
+                        arena.SetPlayerStatsFromLocalPlayer(player, pl, false);
                     }
                     arena.ReadFromStats(player, pl);
                 }