+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Release 1.15.0
+## Arena
+- Added new gamemode: Drown! Kill and survive to buy your escape. Cooperative or Competitive
+- Scoring update: Empty kills and friendly fire now subtract from the user (in Teams, the killer loses the points). Points are no longer granted to everyone else.
+- Fixed a thing that should happen when you win that hidden thing
+- Added Import/Export game settings as code with support for FFA, Team Battle, and Drown
+- Slightly improved Void Amoeba's swim speed
+- Added configurable Void Amoeba lethality. Default multiplier is now 4x. You better run.
+### Modders
+- Added `ExportLocalSettings` and `ImportLocalSettings` virtual functions into `ExternalGameMode` for managing Arena settings 
 # Release 1.14.1
 
 ## Engine